==> 01_Create_Database.ipynb <==
import mysql.connector
from mysql.connector import Error
import os
import ctypes



try:
    conn = mysql.connector.connect(host='localhost',
                                         user='root',
                                         password='root')
    if conn.is_connected():
        #read sql statements
        query=open('SQL_Code_Database_Creation.sql','r').read()
        cursor = conn.cursor(buffered=True)
        results = cursor.execute(query, multi=True)
        '''for cur in results:
            print(type(cur))
            print('cursor:', cur)
            if cur.with_rows:
                print('cursor:', cur) 
                print('result:', cur.fetchall())'''
        
except Error as e:
    print("Error while connecting to MySQL", e)
finally:
    if (conn.is_connected()):
        cursor.close()
        conn.commit()
        conn.close()
        print("MySQL connection is closed")
