==> 02_Extract_Clean_Data.ipynb <==
import pandas as pd
import datetime as dt

## export the data and have an insight of the content

dfinput1 = pd.read_csv(f'./input/input1.csv', low_memory=False)
dfinput2 = pd.read_csv(f'./input/input2.csv', low_memory=False)
dfstore = pd.read_csv(f'./input/Geschaefte.csv', low_memory=False)

dfinput1.head(10)

dfinput2.head(10)

dfinput1.info()

dfinput2.info()

## Add a new attribute to the data sources to keep a  track of the source of the data

dfinput1["Source"] = 1
dfinput1["Source"] = 2

## check whether it exists duplicated rows

dfinput1.loc[dfinput1.duplicated()]

dfinput2.loc[dfinput2.duplicated()]

dfinput1.fillna?

dfinput1.fillna(value={'Open':-1}, inplace=True)

dfinput1['Open']=dfinput1['Open'].astype('int64')

dfinput1.loc[dfinput1['Open'].isnull()]

dfinput1.loc[dfinput1['Open']==-1]

dfinput1.head()

dfinput1.set_index('Id', inplace=True)

## Merge the two datasets related to the sales by day and store

dfinput1['Customers']=-1
dfinput1['Sales']=-1

# create a new column to keep track of the source of the initial data
dfinput1['Source']=1
dfinput2['Source']=2
df = pd.concat([dfinput1, dfinput2], axis=0)
df


df.info()

 df.loc[df['StateHoliday']=='a']

 dfinput1.loc[dfinput1['StateHoliday']=='a']

 dfinput2.loc[dfinput2['StateHoliday']=='a']

df['StateHoliday'].unique()

## adjust data types to the same types in the database to avoid errors when loading the database and export the final transformed data 

df["Date"] = pd.to_datetime(df["Date"])
df["Sales"] = df["Sales"].astype("int32")
df["Customers"] = df["Customers"].astype("int32")

df.to_csv(f"./Cleaned Data/sales.csv")

## load, Explore the file Geschaefte.csv and make transformations when necessary

df=pd.read_csv(f"./input/Geschaefte.csv")
df.info()

df

#Cleaning
df.loc[df["CompetitionDistance"].isna(),"CompetitionDistance"] = -1
df.loc[df["CompetitionOpenSinceMonth"].isna(),"CompetitionOpenSinceMonth"] = -1
df.loc[df["CompetitionOpenSinceYear"].isna(),"CompetitionOpenSinceYear"] = -1
df["CompetitionDistance"] = df["CompetitionDistance"].astype("int32")
df["CompetitionOpenSinceMonth"] = df["CompetitionOpenSinceMonth"].astype("int32")
df["CompetitionOpenSinceYear"] = df["CompetitionOpenSinceYear"].astype("int32")
#df["Promo2SinceWeek"] = df["Promo2SinceWeek"].astype("int64")
#df["Promo2SinceYear"] = df["Promo2SinceYear"].astype("int64")
df['Source']=3;

df.to_csv(f"./Cleaned Data/stores.csv")
