==> 03_Load_Data.ipynb <==
conda install -c anaconda mysql-connector-python

import pandas as pd
import datetime as dt
import hashlib
import mysql.connector

conn = mysql.connector.connect(host='localhost', 
                                         database='rossmann',
                                         user='root',
                                         password='root')
cursor = conn.cursor(buffered=True)


## Feeding informations in the RECORD_SOURCE_REF table

cursor.execute("""delete from RECORD_SOURCE_REF;""")
cursor.execute("""insert into RECORD_SOURCE_REF  (RECORD_SOURCE, RECORD_PATH) VALUES  (1, ".\\input\\input1.csv"),\
(2, ".\\input\\input2.csv") ,(3, ".\\input\\Geschaefte.csv");""")

### Import and load data  of the stores in the appropriate tables

df = pd.read_csv(f'.\Cleaned Data\stores.csv')

for line in range(len(df)):
        
    LOAD_DATETIME=dt.datetime.now()
    RECORD_SOURCE = df.loc[line,'Source']
    STORE_ID = df.loc[line,'Store']
    
    #Create  hashing code of hub store  from load_datetime and business key store_id
    string_hash = str(STORE_ID) + str(RECORD_SOURCE)
    STORE_HSH = hashlib.md5(string_hash.encode()).hexdigest()
    
    # insert the data in the table hub_store
    query = "INSERT INTO hub_store (STORE_HSH, LOAD_DATETIME, RECORD_SOURCE, STORE_ID)\
    VALUES ('" + STORE_HSH + "', '" + str(LOAD_DATETIME) + "', '" + str(RECORD_SOURCE) + "', '" + str(STORE_ID) + "');"
    try:
        cursor.execute(query)
    except mysql.connector.IntegrityError as e:
        print("error code:",e)
   
    #  insert data in the satellite sat_store
    
    store_type = df.loc[line, "StoreType"]
    assortment = df.loc[line, "Assortment"]
    LOAD_END_DATETIME=dt.datetime.now()
    string_hash = str(STORE_HSH) + str(RECORD_SOURCE)+ str(LOAD_END_DATETIME)
    SSTORE_HSH=hashlib.md5(string_hash.encode()).hexdigest()
    
    # generating hashdiff code from attributes of sat_store
    string_hash =  str(store_type) + str(assortment)
    diff_hsh = hashlib.md5(string_hash.encode()).hexdigest()
    
      # insert the data in the satellite table sat_store
    query = "INSERT INTO sat_store (SSTORE_HSH, STORE_HSH, DIFF_HSH, LOAD_DATETIME,LOAD_END_DATETIME, RECORD_SOURCE,\
    STORE_TYPE, ASSORTMENT) VALUES ('" + SSTORE_HSH + "','" + STORE_HSH + "','" + diff_hsh + "', '" + str(LOAD_DATETIME) + "',\
    '" + str(LOAD_END_DATETIME)+ "','" + str(RECORD_SOURCE) + "', '" + str(store_type) + "' , '" + str(assortment) + "');"
    try:
        cursor.execute(query)
    except mysql.connector.IntegrityError as e:
        print("error code:",e)
        
        
   #  insert data in the satellite table sat_promo just when the store applies promo2
   
    promo2 = df.loc[line, "Promo2"]
    if (promo2==1):
    
        promo2_since_week = df.loc[line, "Promo2SinceWeek"] 
        promo2_since_year = df.loc[line, "Promo2SinceYear"] 
        promo_interval = df.loc[line, "PromoInterval"] 
        LOAD_DATETIME=dt.datetime.now()
        LOAD_END_DATETIME=dt.datetime.now()
    # generating hashkey of sat_comp
        string_hash=str(STORE_HSH)+ str(RECORD_SOURCE)+ str(LOAD_END_DATETIME)
        promo_hsh=hashlib.md5(string_hash.encode()).hexdigest()
    
     # generating  hashdiff code from attributes of sat_promo
        string_hash =  str(promo2) + str(promo2_since_week)+ str(promo2_since_year)+str(promo_interval)
        diff_hsh = hashlib.md5(string_hash.encode()).hexdigest()
    
        query = "INSERT INTO sat_promo (PROMO_HSH, STORE_HSH, DIFF_HSH, LOAD_DATETIME,LOAD_END_DATETIME, RECORD_SOURCE,\
         PROMO2, PROMO2_SINCE_WEEK, PROMO2_SINCE_YEAR, PROMO_INTERVAL) VALUES ('" + promo_hsh + "','" + STORE_HSH + "',\
         '" + diff_hsh + "', '" + str(LOAD_DATETIME) + "',  '" + str(LOAD_END_DATETIME)+ "','" + str(RECORD_SOURCE) + "',\
          '" + str(promo2) + "' , '" +str(promo2_since_week) + "'  , '" + str(promo2_since_year) +"' , '" + str( promo_interval) + "');"
        try:
            cursor.execute(query)
        except mysql.connector.IntegrityError as e:
            print("error code:",e)
        
    
    
      #  insert data in the satellite table sat_comp
        
    comp_dist = df.loc[line, "CompetitionDistance"]
    month = df.loc[line, "CompetitionOpenSinceMonth"] 
    year = df.loc[line, "CompetitionOpenSinceYear"] 
    LOAD_DATETIME=dt.datetime.now()
    LOAD_END_DATETIME=dt.datetime.now()
    # generating hashkey of sat_comp
    string_hash=str(STORE_HSH)+ str(RECORD_SOURCE)+ str(LOAD_END_DATETIME)
    comp_hsh=hashlib.md5(string_hash.encode()).hexdigest()
    
     # generating  hashdiff code from attributes of sat_comp
    string_hash =  str(comp_dist) + str(month)+ str(year)
    diff_hsh = hashlib.md5(string_hash.encode()).hexdigest()
    
    query = "INSERT INTO sat_comp (COMP_HSH, STORE_HSH, DIFF_HSH, LOAD_DATETIME,LOAD_END_DATETIME, RECORD_SOURCE,\
   COMP_DISTANCE, COMP_OPEN_MONTH, COMP_OPEN_YEAR) VALUES ('" + comp_hsh + "','" + STORE_HSH + "',\
    '" + diff_hsh + "', '" + str(LOAD_DATETIME) + "',  '" + str(LOAD_END_DATETIME)+ "','" + str(RECORD_SOURCE) + "',\
    '" + str(comp_dist) + "' , '" + str( month) + "', '" + str( year) + "');"
    try:
        cursor.execute(query)
    except mysql.connector.IntegrityError as e:
        print("error code:",e)
    
   

print(cursor.execute(" Select * FROM RECORD_SOURCE_REF;"))
pd.read_sql_query(" Select * FROM sat_comp;", con=conn)

## Load the data from the file sales.csv in the appropriate tables in the datavault model

df = pd.read_csv(f'.\Cleaned Data\sales.csv')

df.head(10)


for line in range(len(df)):
    
    # loading data for hub_day
    date=df.loc[line,'Date']
    record_source=df.loc[line,'Source']
  
    # generating hashkey of hub_day always when it is a new entry
    
    # check if the date already exists to avoid duplicated primary keys in hub_day and sat_day before inserting new entries
    query=cursor.execute("SELECT DAY_HSH FROM hub_day WHERE DATE = '" + str(date) + "';")
    result=cursor.fetchall()
    if not result:
        string_hash=str(date)+ str(record_source)
        day_hsh= hashlib.md5(string_hash.encode()).hexdigest()
        LOAD_DATETIME=dt.datetime.now()
        query = "INSERT INTO hub_day (DAY_HSH, LOAD_DATETIME, RECORD_SOURCE, DATE)\
        VALUES ('" + day_hsh + "', '" + str(LOAD_DATETIME) + "', '" + str(record_source) + "', '" + str(date) + "');"
        try:
            cursor.execute(query)
        except mysql.connector.IntegrityError as e:
            print("error code:",e)
        
         # loading data in the satellite sat_day
            
        day_week= df.loc[line,"DayOfWeek"]
        LOAD_DATETIME=dt.datetime.now()
        LOAD_END_DATETIME=dt.datetime.now()
     # generating hashkey of sat_day
        string_hash=str(day_hsh)+str(record_source)+str(LOAD_END_DATETIME)
        sat_day_hsh= hashlib.md5(string_hash.encode()).hexdigest()
    
      # generating  hashdiff code from attributes of sat_day
        string_hash =  str(day_week) + str(record_source) 
        diff_hsh = hashlib.md5(string_hash.encode()).hexdigest()
    
        query = "INSERT INTO sat_day (SAT_DAY_HSH,DAY_HSH,DIFF_HSH, LOAD_DATETIME,LOAD_END_DATETIME, RECORD_SOURCE, DAY_OF_WEEK)\
             VALUES ('" +sat_day_hsh + "','" + day_hsh + "','" + diff_hsh + "', '" + str(LOAD_DATETIME) + "',\
            '" + str(LOAD_END_DATETIME) + "','" + str(record_source) + "', '" + str(day_week) + "');"
        try:
            cursor.execute(query)
        except mysql.connector.IntegrityError as e:
            print("error code:",e)
    #else: 
    #    print(' This date entry already exists in the hub_day')
        
        
        
    
         
    #loading data in the link link_day
    
    store_id= df.loc[line,'Store']
    store_hsh=pd.read_sql_query("select STORE_HSH from hub_store where STORE_ID= '" + str(store_id)+"';", con=conn).iloc[0,0]
    day_hsh= pd.read_sql_query("select DAY_HSH from hub_day where DATE= '" + str(date)+"';" , con=conn).iloc[0,0]
    
    LOAD_DATETIME=dt.datetime.now()
    # generating hashkey of link_day
    string_hash=str(store_hsh)+ str(day_hsh)+str(record_source)
    link_day_hsh= hashlib.md5(string_hash.encode()).hexdigest()
    
    query = "INSERT INTO link_day (LINK_DAY_HSH,STORE_HSH,DAY_HSH, LOAD_DATETIME, RECORD_SOURCE)\
    VALUES ('" + link_day_hsh + "', '" + store_hsh + "', '" + str(day_hsh) + "', '" + str(LOAD_DATETIME) + "', \
    '" + str(record_source) + "');"
    try:
        cursor.execute(query)
    except mysql.connector.IntegrityError as e:
        print("error code:",e)
 

    #loading data in the satellite  sat_day_facts
    opens=df.loc[line,'Open']
    promo=df.loc[line,'Promo']
    state_hol=df.loc[line,'StateHoliday']
    school_hol=df.loc[line,'SchoolHoliday']
    sales=df.loc[line,'Sales']
    customers=df.loc[line,'Customers']
    LOAD_DATETIME=dt.datetime.now()
    LOAD_END_DATETIME=dt.datetime.now()
    
     # generating hashkey of sat_day_facts
    string_hash=str(link_day_hsh)+str(record_source)+ str(LOAD_END_DATETIME)
    sat_dayf_hsh= hashlib.md5(string_hash.encode()).hexdigest()
    
     # generating  hashdiff code from attributes of sat_day_facts
    string_hash =  str(opens) + str(state_hol) + str(school_hol) + str(promo) + str(customers) + str(sales)
    diff_hsh = hashlib.md5(string_hash.encode()).hexdigest()
    
    query = "INSERT INTO sat_day_facts (SAT_DAYF_HSH, LINK_DAY_HSH, DIFF_HSH, LOAD_DATETIME,LOAD_END_DATETIME, RECORD_SOURCE,\
   OPEN, STATE_HOLIDAY, SCHOOL_HOLIDAY, PROMO, CUSTOMERS, SALES) VALUES ('" + sat_dayf_hsh + "','" + link_day_hsh + "',\
    '" + diff_hsh + "', '" + str(LOAD_DATETIME) + "',  '" + str(LOAD_END_DATETIME)+ "','" + str(RECORD_SOURCE) + "',\
    '" + str(opens) + "' , '" + str( state_hol) + "', '" + str( school_hol) + "' , '" + str( promo) + "',\
      '" + str( customers) + "', '" + str( sales) + "');"
    try:
        cursor.execute(query)
    except mysql.connector.IntegrityError as e:
        print("error code:",e)
    
  

   

pd.read_sql_query("select * from hub_day;" , con=conn)

pd.read_sql_query("select * from sat_day_facts LIMIT 20;" , con=conn)

cursor.close()
conn.commit()
conn.close()
