==> 04_Load_business_data.ipynb <==
import pandas as pd
import datetime as dt
import hashlib
import mysql.connector

conn = mysql.connector.connect(host='localhost',
                                         database='rossmann',
                                         user='root',
                                         password='root')
cursor = conn.cursor(buffered=True)


query= """ insert into sat_day_report (DAY_HSH,SALES_COUNT,CUSTOMER_COUNT)
        select h.DAY_HSH,  SUM(SALES), SUM(CUSTOMERS) FROM sat_day_facts s inner join link_day l 
        on s.LINK_DAY_HSH=l.LINK_DAY_HSH inner join hub_day  h  on h.DAY_HSH=l.DAY_HSH
        group by h.DAY_HSH; """ 

try:
    cursor.execute(query)
except mysql.connector.IntegrityError as e:
        print("error code:",e)

pd.read_sql_query(" Select * FROM sat_day_report;", con=conn).head(10)

select 
    t.storeid as storeid,
    t.artid as artid,
    t.totamount,
    t.hour,
    (
        select 1 + count(*)
        from t t1
        where 
            t1.storeid = t.storeid 
            and t1.artid = t.artid
            and t1.hour > t.hour

    ) as rn
from t
order by t.storeid, t.artid

WITH t AS
 (select h.STORE_HSH,  AVG(s.SALES) as SALES_AVG , AVG(s.CUSTOMERS) as CSTMR_AVG
 FROM sat_day_facts s inner join link_day l
  on s.LINK_DAY_HSH=l.LINK_DAY_HSH inner join hub_store  h  on h.STORE_HSH=l.STORE_HSH
  group by h.STORE_HSH)

  select t.STORE_HSH, NOW() as LOAD_DATETIME, NOW() as LOAD_END_DATETIME, t.SALES_AVG, t.CSTMR_AVG,
  (  select 1 + count(*)
      from t t1
      where  t1.SALES_AVG > t.SALES_AVG) as SALES_RANK;


query= """   insert into sat_store_biz (STORE_HSH, LOAD_DATETIME, LOAD_END_DATETIME,SALES_AVG, CSTMR_AVG, SALES_RANK)
              WITH t as  (select h.STORE_HSH,  AVG(s.SALES) as SALES_AVG , AVG(s.CUSTOMERS) as CSTMR_AVG 
              FROM sat_day_facts s inner join link_day l 
              on s.LINK_DAY_HSH=l.LINK_DAY_HSH inner join hub_store  h  on h.STORE_HSH=l.STORE_HSH 
                  group by h.STORE_HSH) 
           select t.STORE_HSH, NOW() as LOAD_DATETIME, NOW() as LOAD_END_DATETIME, t.SALES_AVG, t.CSTMR_AVG,
          (  select 1 + count(*)
            from t t1
            where  t1.SALES_AVG > t.SALES_AVG
           ) as SALES_RANK
           from t
        
            ; """ 

try:
    cursor.execute(query)
except mysql.connector.IntegrityError as e:
        print("error code:",e)

cursor.close()
conn.commit()
conn.close()
